+++ src/introspect/controller.ts
@@ -70,7 +70,8 @@ class IntrospectionController extends Controller {
     }
     if (foundToken) {
       const privileges = (await privilegeService.get(foundToken.principal)).getAll();
-      const client = await oauth2ClientService.findById(foundToken.clientId);
+      // For 'developer tokens' clientId will be 0.
+      const client = foundToken.clientId !== 0 ? await oauth2ClientService.findById(foundToken.clientId) : undefined;
 
       switch (foundTokenType!) {